Refactor tables and functions to components; add tests
